requirements/requirements-docs.txt
#
# This file is autogenerated by hatch-pip-compile with Python 3.11
#
# [constraints] requirements/requirements.txt (SHA256: e01544cf72f8ac579876c9edae8879cd43af1c56a495cd338dcbe0b69ae04ded)
#
# - nodeenv
# - cython
# - anyio
# - coverage[toml]>=6.2
# - pytest
# - pytest-cov
# - pytest-mock
# - pytest-vcr
# - pytest-sugar
# - pytest-click
# - pytest-xdist
# - sphinx>=7.1.2
# - sphinx-autobuild>=2021.3.14
# - sphinx-copybutton>=0.5.2
# - litestar-sphinx-theme@ git+https://github.com/litestar-org/litestar-sphinx-theme.git
# - sphinx-click>=5.0.1
# - sphinx-toolbox>=3.5.0
# - sphinx-design>=0.5.0
# - sphinxcontrib-mermaid>=0.9.2
# - auto-pytabs[sphinx]>=0.4.0
# - pytest
# - psycopg
# - elasticsearch7[async]
# - elasticsearch8[async]
# - aioodbc
# - asyncmy>=0.2.9
# - oracledb
# - asyncpg>=0.29.0
# - redis
# - google-cloud-spanner
#

accessible-pygments==0.0.4
    # via pydata-sphinx-theme
aiohttp==3.9.3
    # via
    #   elasticsearch7
    #   elasticsearch8
aioodbc==0.5.0
aiosignal==1.3.1
    # via aiohttp
alabaster==0.7.16
    # via sphinx
anyio==4.3.0
apeye==1.4.1
    # via sphinx-toolbox
apeye-core==1.1.5
    # via apeye
async-timeout==4.0.3
    # via asyncpg
asyncmy==0.2.9
asyncpg==0.29.0
attrs==23.2.0
    # via aiohttp
auto-pytabs==0.4.0
autodocsumm==0.2.12
    # via sphinx-toolbox
babel==2.14.0
    # via
    #   pydata-sphinx-theme
    #   sphinx
beautifulsoup4==4.12.3
    # via
    #   pydata-sphinx-theme
    #   sphinx-toolbox
cachecontrol==0.14.0
    # via sphinx-toolbox
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   elastic-transport
    #   elasticsearch7
    #   requests
cffi==1.16.0
    # via cryptography
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   pytest-click
    #   sphinx-click
colorama==0.4.6
    # via sphinx-autobuild
coverage==7.4.4
    # via pytest-cov
cryptography==42.0.5
    # via oracledb
cssutils==2.10.2
    # via dict2css
cython==3.0.10
deprecated==1.2.14
    # via google-cloud-spanner
dict2css==0.3.0.post1
    # via sphinx-toolbox
docutils==0.20.1
    # via
    #   pydata-sphinx-theme
    #   sphinx
    #   sphinx-click
    #   sphinx-prompt
    #   sphinx-tabs
    #   sphinx-toolbox
domdf-python-tools==3.8.0.post2
    # via
    #   apeye
    #   apeye-core
    #   dict2css
    #   sphinx-toolbox
elastic-transport==8.13.0
    # via elasticsearch8
elasticsearch7==7.17.9
elasticsearch8==8.13.0
execnet==2.0.2
    # via pytest-xdist
filelock==3.13.3
    # via
    #   cachecontrol
    #   sphinx-toolbox
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core==2.18.0
    # via
    #   google-cloud-core
    #   google-cloud-spanner
google-auth==2.29.0
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-core==2.4.1
    # via google-cloud-spanner
google-cloud-spanner==3.44.0
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-spanner
grpc-interceptor==0.15.4
    # via google-cloud-spanner
grpcio==1.62.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpc-interceptor
    #   grpcio-status
grpcio-status==1.62.1
    # via google-api-core
html5lib==1.1
    # via sphinx-toolbox
idna==3.6
    # via
    #   anyio
    #   apeye-core
    #   requests
    #   yarl
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
jinja2==3.1.3
    # via
    #   sphinx
    #   sphinx-jinja2-compat
litestar-sphinx-theme @ git+https://github.com/litestar-org/litestar-sphinx-theme.git@c5ce66aadc8f910c24f54bf0d172798c237a67eb
livereload==2.6.3
    # via sphinx-autobuild
markupsafe==2.1.5
    # via
    #   jinja2
    #   sphinx-jinja2-compat
msgpack==1.0.8
    # via cachecontrol
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
natsort==8.4.0
    # via domdf-python-tools
nodeenv==1.8.0
oracledb==2.1.1
packaging==24.0
    # via
    #   pydata-sphinx-theme
    #   pytest
    #   pytest-sugar
    #   sphinx
platformdirs==4.2.0
    # via apeye
pluggy==1.4.0
    # via pytest
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-spanner
protobuf==4.25.3
    # via
    #   google-api-core
    #   google-cloud-spanner
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg==3.1.18
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pycparser==2.22
    # via cffi
pydata-sphinx-theme==0.15.2
    # via litestar-sphinx-theme
pygments==2.17.2
    # via
    #   accessible-pygments
    #   pydata-sphinx-theme
    #   sphinx
    #   sphinx-prompt
    #   sphinx-tabs
pyodbc==5.1.0
    # via aioodbc
pytest==8.1.1
    # via
    #   pytest-click
    #   pytest-cov
    #   pytest-mock
    #   pytest-sugar
    #   pytest-vcr
    #   pytest-xdist
pytest-click==1.1.0
pytest-cov==5.0.0
pytest-mock==3.14.0
pytest-sugar==1.0.0
pytest-vcr==1.0.2
pytest-xdist==3.5.0
pyyaml==6.0.1
    # via vcrpy
redis==5.0.3
requests==2.31.0
    # via
    #   apeye
    #   cachecontrol
    #   google-api-core
    #   sphinx
rsa==4.9
    # via google-auth
ruamel-yaml==0.18.6
    # via sphinx-toolbox
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
ruff==0.3.5
    # via auto-pytabs
setuptools==69.2.0
    # via nodeenv
six==1.16.0
    # via
    #   html5lib
    #   livereload
sniffio==1.3.1
    # via anyio
snowballstemmer==2.2.0
    # via sphinx
soupsieve==2.5
    # via beautifulsoup4
sphinx==7.2.6
    # via
    #   auto-pytabs
    #   autodocsumm
    #   pydata-sphinx-theme
    #   sphinx-autobuild
    #   sphinx-autodoc-typehints
    #   sphinx-click
    #   sphinx-copybutton
    #   sphinx-design
    #   sphinx-prompt
    #   sphinx-tabs
    #   sphinx-toolbox
sphinx-autobuild==2024.2.4
sphinx-autodoc-typehints==2.0.0
    # via sphinx-toolbox
sphinx-click==5.1.0
sphinx-copybutton==0.5.2
sphinx-design==0.5.0
    # via litestar-sphinx-theme
sphinx-jinja2-compat==0.2.0.post1
    # via sphinx-toolbox
sphinx-prompt==1.8.0
    # via sphinx-toolbox
sphinx-tabs==3.4.5
    # via sphinx-toolbox
sphinx-toolbox==3.5.0
sphinxcontrib-applehelp==1.0.8
    # via sphinx
sphinxcontrib-devhelp==1.0.6
    # via sphinx
sphinxcontrib-htmlhelp==2.0.5
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-mermaid==0.9.2
sphinxcontrib-qthelp==1.0.7
    # via sphinx
sphinxcontrib-serializinghtml==1.1.10
    # via sphinx
sqlparse==0.4.4
    # via google-cloud-spanner
tabulate==0.9.0
    # via sphinx-toolbox
termcolor==2.4.0
    # via pytest-sugar
tornado==6.4
    # via livereload
typing-extensions==4.10.0
    # via
    #   domdf-python-tools
    #   psycopg
    #   pydata-sphinx-theme
    #   sphinx-toolbox
urllib3==1.26.18
    # via
    #   elastic-transport
    #   elasticsearch7
    #   requests
vcrpy==6.0.1
    # via pytest-vcr
webencodings==0.5.1
    # via html5lib
wrapt==1.16.0
    # via
    #   deprecated
    #   vcrpy
yarl==1.9.4
    # via
    #   aiohttp
    #   vcrpy
